requirements.txt
crispy-bootstrap4==2023.1
Django==5.0
django-braces==1.15.0
django-crispy-forms==2.1
django-filter==23.5
django-model-utils==4.3.1
Pillow==10.1.0
psycopg==3.1.16
